stapbpf: Fix BPF interpreter regression from Clang compatibility changes
The commit 131d504a44f changed SystemTap-specific BPF helper IDs from
const variables to an enum. In stapbpf/bpfinterp.cxx, the interpreter
was comparing sign-extended 64-bit values against these enum constants.
For negative IDs like BPF_FUNC_map_get_next_key, the comparison failed,
triggering an "unknown helper function" abort.

This fix casts the helper ID to uint32_t in the switch statement to
ensure correct matching for both standard and SystemTap-specific IDs.

Related to PR34283 (clang build compatibility tweaks).

+++ stapbpf/bpfinterp.cxx
@@ -1039,7 +1039,7 @@ bpf_interpret(size_t ninsns, const struct bpf_insn insns[],
 	  continue;
 
 	case BPF_JMP | BPF_CALL:
-	  switch (si)
+	  switch ((uint32_t)si)
 	    {
 	    case BPF_FUNC_map_lookup_elem:
 	      {
